+++ ecal/core/src/v5/service/ecal_service_client_impl.cpp
@@ -113,7 +113,7 @@ namespace eCAL
             event_data.time = data_.time;
             event_data.attr.hname = service_id_.service_id.host_name;
             event_data.attr.sname = service_id_.service_name;
-            event_data.attr.pid   = service_id_.service_id.process_id;
+            event_data.attr.pid   = static_cast<int32_t>(service_id_.service_id.process_id);
             event_data.attr.sid   = service_id_.service_id.entity_id;
             callback->second(service_id_.service_name.c_str(), &event_data);
           }
